entitycaps: Move XEP-115-related things into own file
