+++ tests/entitycaps/test_service.py
@@ -31,7 +31,6 @@ import aioxmpp.disco as disco
 import aioxmpp.service as service
 import aioxmpp.stanza as stanza
 import aioxmpp.structs as structs
-import aioxmpp.forms.xso as forms_xso
 import aioxmpp.xml
 
 import aioxmpp.entitycaps.service as entitycaps_service
@@ -45,543 +44,6 @@ from aioxmpp.testutils import (
 
 
 TEST_FROM = structs.JID.fromstr("foo@bar.example/r1")
-
-
-class Testbuild_identities_string(unittest.TestCase):
-    def test_identities(self):
-        identities = [
-            disco.xso.Identity(category="fnord",
-                               type_="bar"),
-            disco.xso.Identity(category="client",
-                               type_="bot",
-                               name="aioxmpp library"),
-            disco.xso.Identity(category="client",
-                               type_="bot",
-                               name="aioxmpp Bibliothek",
-                               lang=structs.LanguageTag.fromstr("de-de")),
-        ]
-
-        self.assertEqual(
-            b"client/bot//aioxmpp library<"
-            b"client/bot/de-de/aioxmpp Bibliothek<"
-            b"fnord/bar//<",
-            entitycaps_service.build_identities_string(identities)
-        )
-
-    def test_escaping(self):
-        identities = [
-            disco.xso.Identity(category="fnord",
-                               type_="bar"),
-            disco.xso.Identity(category="client",
-                               type_="bot",
-                               name="aioxmpp library > 0.5"),
-            disco.xso.Identity(category="client",
-                               type_="bot",
-                               name="aioxmpp Bibliothek <& 0.5",
-                               lang=structs.LanguageTag.fromstr("de-de")),
-        ]
-
-        self.assertEqual(
-            b"client/bot//aioxmpp library &gt; 0.5<"
-            b"client/bot/de-de/aioxmpp Bibliothek &lt;&amp; 0.5<"
-            b"fnord/bar//<",
-            entitycaps_service.build_identities_string(identities)
-        )
-
-    def test_reject_duplicate_identities(self):
-        identities = [
-            disco.xso.Identity(category="fnord",
-                               type_="bar"),
-            disco.xso.Identity(category="client",
-                               type_="bot",
-                               name="aioxmpp library > 0.5"),
-            disco.xso.Identity(category="client",
-                               type_="bot",
-                               name="aioxmpp Bibliothek <& 0.5",
-                               lang=structs.LanguageTag.fromstr("de-de")),
-            disco.xso.Identity(category="client",
-                               type_="bot",
-                               name="aioxmpp library > 0.5"),
-        ]
-
-        with self.assertRaisesRegex(ValueError,
-                                    "duplicate identity"):
-            entitycaps_service.build_identities_string(identities)
-
-
-class Testbuild_features_string(unittest.TestCase):
-    def test_features(self):
-        features = [
-            "http://jabber.org/protocol/disco#info",
-            "http://jabber.org/protocol/caps",
-            "http://jabber.org/protocol/disco#items",
-        ]
-
-        self.assertEqual(
-            b"http://jabber.org/protocol/caps<"
-            b"http://jabber.org/protocol/disco#info<"
-            b"http://jabber.org/protocol/disco#items<",
-            entitycaps_service.build_features_string(features)
-        )
-
-    def test_escaping(self):
-        features = [
-            "http://jabber.org/protocol/c<>&aps",
-        ]
-
-        self.assertEqual(
-            b"http://jabber.org/protocol/c&lt;&gt;&amp;aps<",
-            entitycaps_service.build_features_string(features)
-        )
-
-    def test_reject_duplicate_features(self):
-        features = [
-            "http://jabber.org/protocol/disco#info",
-            "http://jabber.org/protocol/caps",
-            "http://jabber.org/protocol/disco#items",
-            "http://jabber.org/protocol/caps",
-        ]
-
-        with self.assertRaisesRegex(ValueError,
-                                    "duplicate feature"):
-            entitycaps_service.build_features_string(features)
-
-
-class Testbuild_forms_string(unittest.TestCase):
-    def test_xep_form(self):
-        forms = [forms_xso.Data(type_=forms_xso.DataType.FORM)]
-        forms[0].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "urn:xmpp:dataforms:softwareinfo",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os_version",
-                values=[
-                    "10.5.1",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os",
-                values=[
-                    "Mac",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="ip_version",
-                values=[
-                    "ipv4",
-                    "ipv6",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="software",
-                values=[
-                    "Psi",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="software_version",
-                values=[
-                    "0.11",
-                ]
-            ),
-        ])
-
-        self.assertEqual(
-            b"urn:xmpp:dataforms:softwareinfo<"
-            b"ip_version<ipv4<ipv6<"
-            b"os<Mac<"
-            b"os_version<10.5.1<"
-            b"software<Psi<"
-            b"software_version<0.11<",
-            entitycaps_service.build_forms_string(forms)
-        )
-
-    def test_value_and_var_escaping(self):
-        forms = [forms_xso.Data(type_=forms_xso.DataType.FORM)]
-        forms[0].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "urn:xmpp:<dataforms:softwareinfo",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os_version",
-                values=[
-                    "10.&5.1",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os>",
-                values=[
-                    "Mac",
-                ]
-            ),
-        ])
-
-        self.assertEqual(
-            b"urn:xmpp:&lt;dataforms:softwareinfo<"
-            b"os&gt;<Mac<"
-            b"os_version<10.&amp;5.1<",
-            entitycaps_service.build_forms_string(forms)
-        )
-
-    def test_reject_multiple_identical_form_types(self):
-        forms = [forms_xso.Data(type_=forms_xso.DataType.FORM), forms_xso.Data(type_=forms_xso.DataType.FORM)]
-
-        forms[0].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "urn:xmpp:dataforms:softwareinfo",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os_version",
-                values=[
-                    "10.5.1",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os",
-                values=[
-                    "Mac",
-                ]
-            ),
-        ])
-
-        forms[1].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "urn:xmpp:dataforms:softwareinfo",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os_version",
-                values=[
-                    "10.5.1",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os",
-                values=[
-                    "Mac",
-                ]
-            ),
-        ])
-
-        with self.assertRaisesRegex(
-                ValueError,
-                "multiple forms of type b'urn:xmpp:dataforms:softwareinfo'"):
-            entitycaps_service.build_forms_string(forms)
-
-    def test_reject_form_with_multiple_different_types(self):
-        forms = [forms_xso.Data(type_=forms_xso.DataType.FORM)]
-
-        forms[0].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "urn:xmpp:dataforms:softwareinfo",
-                    "urn:xmpp:dataforms:softwarefoo",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os_version",
-                values=[
-                    "10.5.1",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os",
-                values=[
-                    "Mac",
-                ]
-            ),
-        ])
-
-        with self.assertRaisesRegex(
-                ValueError,
-                "form with multiple types"):
-            entitycaps_service.build_forms_string(forms)
-
-    def test_ignore_form_without_type(self):
-        forms = [forms_xso.Data(type_=forms_xso.DataType.FORM),
-                 forms_xso.Data(type_=forms_xso.DataType.FORM)]
-
-        forms[0].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os_version",
-                values=[
-                    "10.5.1",
-                ]
-            ),
-        ])
-
-        forms[1].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "urn:xmpp:dataforms:softwareinfo",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os_version",
-                values=[
-                    "10.5.1",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os",
-                values=[
-                    "Mac",
-                ]
-            ),
-        ])
-
-        self.assertEqual(
-            b"urn:xmpp:dataforms:softwareinfo<"
-            b"os<Mac<"
-            b"os_version<10.5.1<",
-            entitycaps_service.build_forms_string(forms)
-        )
-
-    def test_accept_form_with_multiple_identical_types(self):
-        forms = [forms_xso.Data(type_=forms_xso.DataType.FORM)]
-
-        forms[0].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "urn:xmpp:dataforms:softwareinfo",
-                    "urn:xmpp:dataforms:softwareinfo",
-                ]
-            ),
-
-            forms_xso.Field(
-                var="os_version",
-                values=[
-                    "10.5.1",
-                ]
-            ),
-        ])
-
-        entitycaps_service.build_forms_string(forms)
-
-    def test_multiple(self):
-        forms = [forms_xso.Data(type_=forms_xso.DataType.FORM), forms_xso.Data(type_=forms_xso.DataType.FORM)]
-
-        forms[0].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "uri:foo",
-                ]
-            ),
-        ])
-
-        forms[1].fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "uri:bar",
-                ]
-            ),
-        ])
-
-        self.assertEqual(
-            b"uri:bar<uri:foo<",
-            entitycaps_service.build_forms_string(forms)
-        )
-
-
-class Testhash_query(unittest.TestCase):
-    def test_impl(self):
-        self.maxDiff = None
-        base = unittest.mock.Mock()
-
-        with contextlib.ExitStack() as stack:
-            stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.build_identities_string",
-                new=base.build_identities_string,
-            ))
-
-            stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.build_features_string",
-                new=base.build_features_string,
-            ))
-
-            stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.build_forms_string",
-                new=base.build_forms_string,
-            ))
-
-            stack.enter_context(unittest.mock.patch(
-                "hashlib.new",
-                new=base.hashlib_new,
-            ))
-
-            stack.enter_context(unittest.mock.patch(
-                "base64.b64encode",
-                new=base.base64_b64encode,
-            ))
-
-            result = entitycaps_service.hash_query(
-                base.query,
-                base.algo
-            )
-
-        calls = list(base.mock_calls)
-        self.assertSequenceEqual(
-            calls,
-            [
-                unittest.mock.call.hashlib_new(base.algo),
-                unittest.mock.call.build_identities_string(
-                    base.query.identities,
-                ),
-                unittest.mock.call.hashlib_new().update(
-                    base.build_identities_string()
-                ),
-                unittest.mock.call.build_features_string(
-                    base.query.features
-                ),
-                unittest.mock.call.hashlib_new().update(
-                    base.build_features_string()
-                ),
-                unittest.mock.call.build_forms_string(
-                    base.query.exts
-                ),
-                unittest.mock.call.hashlib_new().update(
-                    base.build_forms_string()
-                ),
-                unittest.mock.call.hashlib_new().digest(),
-                unittest.mock.call.base64_b64encode(
-                    base.hashlib_new().digest()
-                ),
-                unittest.mock.call.base64_b64encode().decode("ascii")
-            ]
-        )
-
-        self.assertEqual(
-            result,
-            base.base64_b64encode().decode()
-        )
-
-    def test_simple_xep_data(self):
-        info = disco.xso.InfoQuery()
-        info.identities.extend([
-            disco.xso.Identity(category="client",
-                               name="Exodus 0.9.1",
-                               type_="pc"),
-        ])
-
-        info.features.update({
-                "http://jabber.org/protocol/caps",
-                "http://jabber.org/protocol/disco#info",
-                "http://jabber.org/protocol/disco#items",
-                "http://jabber.org/protocol/muc",
-        })
-
-        self.assertEqual(
-            "QgayPKawpkPSDYmwT/WM94uAlu0=",
-            entitycaps_service.hash_query(info, "sha1")
-        )
-
-    def test_complex_xep_data(self):
-        info = disco.xso.InfoQuery()
-        info.identities.extend([
-            disco.xso.Identity(category="client",
-                               name="Psi 0.11",
-                               type_="pc",
-                               lang=structs.LanguageTag.fromstr("en")),
-            disco.xso.Identity(category="client",
-                               name="Ψ 0.11",
-                               type_="pc",
-                               lang=structs.LanguageTag.fromstr("el")),
-        ])
-
-        info.features.update({
-            "http://jabber.org/protocol/caps",
-            "http://jabber.org/protocol/disco#info",
-            "http://jabber.org/protocol/disco#items",
-            "http://jabber.org/protocol/muc",
-        })
-
-        ext_form = forms_xso.Data(type_=forms_xso.DataType.FORM)
-
-        ext_form.fields.extend([
-            forms_xso.Field(
-                var="FORM_TYPE",
-                values=[
-                    "urn:xmpp:dataforms:softwareinfo"
-                ]
-            ),
-
-            forms_xso.Field(
-                var="ip_version",
-                values=[
-                    "ipv6",
-                    "ipv4",
-                ]
-            ),
-            forms_xso.Field(
-                var="os",
-                values=[
-                    "Mac"
-                ]
-            ),
-            forms_xso.Field(
-                var="os_version",
-                values=[
-                    "10.5.1",
-                ]
-            ),
-            forms_xso.Field(
-                var="software",
-                values=[
-                    "Psi",
-                ]
-            ),
-            forms_xso.Field(
-                var="software_version",
-                values=[
-                    "0.11"
-                ]
-            ),
-        ])
-
-        info.exts.append(ext_form)
-
-        self.assertEqual(
-            "q07IKJEyjvHSyhy//CH0CxmKi8w=",
-            entitycaps_service.hash_query(info, "sha1")
-        )
 
 
 _src = io.BytesIO(b"""\
@@ -1253,7 +715,7 @@ class TestService(unittest.TestCase):
         base.disco.query_info.side_effect = None
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.hash_query",
+                "aioxmpp.entitycaps.caps115.hash_query",
                 new=base.hash_query
             ))
 
@@ -1311,7 +773,7 @@ class TestService(unittest.TestCase):
         base.disco.query_info.side_effect = None
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.hash_query",
+                "aioxmpp.entitycaps.caps115.hash_query",
                 new=base.hash_query
             ))
 
@@ -1365,7 +827,7 @@ class TestService(unittest.TestCase):
         base.disco.query_info.side_effect = None
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.hash_query",
+                "aioxmpp.entitycaps.caps115.hash_query",
                 new=base.hash_query
             ))
 
@@ -1519,7 +981,7 @@ class TestService(unittest.TestCase):
         base = unittest.mock.Mock()
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.hash_query",
+                "aioxmpp.entitycaps.caps115.hash_query",
                 new=base.hash_query
             ))
 
@@ -1602,7 +1064,7 @@ class TestService(unittest.TestCase):
         base = unittest.mock.Mock()
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.hash_query",
+                "aioxmpp.entitycaps.caps115.hash_query",
                 new=base.hash_query
             ))
             base.hash_query.return_value = "hash_query_result"
@@ -1625,7 +1087,7 @@ class TestService(unittest.TestCase):
         base = unittest.mock.Mock()
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.hash_query",
+                "aioxmpp.entitycaps.caps115.hash_query",
                 new=base.hash_query
             ))
             base.hash_query.return_value = "hash_query_result"
@@ -1651,7 +1113,7 @@ class TestService(unittest.TestCase):
         base = unittest.mock.Mock()
         with contextlib.ExitStack() as stack:
             stack.enter_context(unittest.mock.patch(
-                "aioxmpp.entitycaps.service.hash_query",
+                "aioxmpp.entitycaps.caps115.hash_query",
                 new=base.hash_query
             ))
             base.hash_query.return_value = "hash_query_result"
